Restore the `Commands` namespace and make better use of it
