--- lib/packwerk/cli/lazy_loaded_command.rb
@@ -1,65 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-module Packwerk
-  class Cli
-    class LazyLoadedCommand
-      extend T::Sig
-
-      class << self
-        extend T::Sig
-
-        sig { params(name: String, aliases: T::Array[String]).void }
-        def register(name, aliases: [])
-          registry << new(name, aliases: aliases)
-        end
-
-        sig { params(name_or_alias: String).returns(T.nilable(T.class_of(Cli::BaseCommand))) }
-        def class_for(name_or_alias)
-          registry
-            .find { |command| command.matches_command?(name_or_alias) }
-            &.command_class
-        end
-
-        sig { returns(T::Array[LazyLoadedCommand]) }
-        def all
-          registry.dup
-        end
-
-        private
-
-        sig { returns(T::Array[LazyLoadedCommand]) }
-        def registry
-          @registry ||= T.let([], T.nilable(T::Array[LazyLoadedCommand]))
-        end
-      end
-
-      sig { returns(String) }
-      attr_reader :name
-
-      sig { params(name: String, aliases: T::Array[String]).void }
-      def initialize(name, aliases: [])
-        @name = name
-        @aliases = aliases
-      end
-
-      sig { returns(T.class_of(Cli::BaseCommand)) }
-      def command_class
-        classname = @name.sub(" ", "_").underscore.classify + "Command"
-        Cli.const_get(classname) # rubocop:disable Sorbet/ConstantsFromStrings
-      end
-
-      sig { returns(String) }
-      def description
-        command_class.description
-      end
-
-      sig { params(name_or_alias: String).returns(T::Boolean) }
-      def matches_command?(name_or_alias)
-        @name == name_or_alias || @aliases.include?(name_or_alias)
-      end
-    end
-
-    private_constant :LazyLoadedCommand
-  end
-end
