+++ lib/packwerk/commands/uses_parse_run.rb
@@ -4,12 +4,12 @@
 require "optparse"
 
 module Packwerk
-  class Cli
+  module Commands
     module UsesParseRun
       extend T::Sig
       extend T::Helpers
 
-      requires_ancestor { Cli::BaseCommand }
+      requires_ancestor { BaseCommand }
 
       sig do
         params(
